Move ffmpeg cache redirect out of Engine into ArtifactManager app-init handler
