+++ tests/unit/retained_mode/managers/test_artifact_manager.py
@@ -8,10 +8,14 @@ from unittest.mock import Mock
 
 import anyio
 import pytest
+import static_ffmpeg.run
 from PIL import Image
 from pydantic import ValidationError
+from xdg_base_dirs import xdg_data_home
 
 from griptape_nodes.common.macro_parser import ParsedMacro
+from griptape_nodes.retained_mode.engine import Engine
+from griptape_nodes.retained_mode.events.app_events import AppInitializationComplete
 from griptape_nodes.retained_mode.events.artifact_events import (
     GeneratePreviewRequest,
     GeneratePreviewResultFailure,
@@ -38,6 +42,7 @@ from griptape_nodes.retained_mode.managers.artifact_providers import (
     BaseArtifactProvider,
     ImageArtifactProvider,
 )
+from griptape_nodes.utils import ffmpeg_cache
 
 if TYPE_CHECKING:
     from griptape_nodes.retained_mode.managers.artifact_providers.image.preview_generators.pil_thumbnail_generator import (
@@ -1688,3 +1693,66 @@ class TestProviderRegistrationConfigLogLevels:
         assert error_records == [], (
             f"Provider registration produced ERROR-level logs: {[r.message for r in error_records]}"
         )
+
+
+class TestFfmpegCacheRedirect:
+    """App initialization must move ffmpeg off `static_ffmpeg`'s own package directory.
+
+    That directory is read-only when the engine runs from a packaged app -- notably the Linux
+    AppImage's FUSE mount, where it made every ffmpeg-dependent node fail with Errno 30. The
+    redirect lives here rather than in engine boot because the video artifact provider is what
+    depends on `static_ffmpeg`. It is process-wide and installed once, so `_redirect_installed`
+    is forced to `False` here to make each test's broadcast act as the process's first.
+    """
+
+    @pytest.fixture(autouse=True)
+    def _restore_static_ffmpeg_globals(self, monkeypatch: pytest.MonkeyPatch) -> None:
+        monkeypatch.setattr(static_ffmpeg.run, "SELF_DIR", static_ffmpeg.run.SELF_DIR)
+        monkeypatch.setattr(static_ffmpeg.run, "LOCK_FILE", static_ffmpeg.run.LOCK_FILE)
+        monkeypatch.setattr(ffmpeg_cache, "_redirect_installed", False)
+
+    @pytest.mark.asyncio
+    async def test_redirects_away_from_the_package_directory(self) -> None:
+        package_dir = Path(static_ffmpeg.run.__file__).parent
+        engine = Engine()
+
+        await engine.artifact_manager.on_app_initialization_complete(AppInitializationComplete())
+
+        assert not Path(static_ffmpeg.run.SELF_DIR).is_relative_to(package_dir)
+
+    @pytest.mark.asyncio
+    async def test_defaults_to_xdg_data_home(self) -> None:
+        engine = Engine()
+
+        await engine.artifact_manager.on_app_initialization_complete(AppInitializationComplete())
+
+        assert Path(static_ffmpeg.run.SELF_DIR) == xdg_data_home() / "griptape_nodes" / "ffmpeg"
+
+    @pytest.mark.asyncio
+    async def test_honors_config_env_var(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        """`GTN_CONFIG_FFMPEG_DIRECTORY` is the hook a packaged app uses to supply its own binaries."""
+        monkeypatch.setenv("GTN_CONFIG_FFMPEG_DIRECTORY", str(tmp_path))
+        engine = Engine()
+
+        await engine.artifact_manager.on_app_initialization_complete(AppInitializationComplete())
+
+        assert Path(static_ffmpeg.run.SELF_DIR) == tmp_path
+
+    @pytest.mark.asyncio
+    async def test_second_broadcast_leaves_the_redirect_alone(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+    ) -> None:
+        """The redirect is per process, not per engine object or per broadcast.
+
+        Later engines and re-broadcasts must not move the cache out from under callers
+        that already resolved ffmpeg through the first install.
+        """
+        monkeypatch.setenv("GTN_CONFIG_FFMPEG_DIRECTORY", str(tmp_path / "first"))
+        first_engine = Engine()
+        await first_engine.artifact_manager.on_app_initialization_complete(AppInitializationComplete())
+
+        monkeypatch.setenv("GTN_CONFIG_FFMPEG_DIRECTORY", str(tmp_path / "second"))
+        second_engine = Engine()
+        await second_engine.artifact_manager.on_app_initialization_complete(AppInitializationComplete())
+
+        assert Path(static_ffmpeg.run.SELF_DIR) == tmp_path / "first"
